# perf: trace.go 高性能优化 - 最大7.1x性能提升，零内存分配