--- a/.gitignore
+++ b/.gitignore
@@ -35,3 +35,5 @@ Thumbs.db
 # Temporary directories
 tmp/
 temp/
+
+*.prof