… cpu.prof
… go.sum
… mem.prof